[MOV] autofill: move autofill stores in correct folder
Task: 6395361
