+++ tests/collaborative/collaborative_session_squish.test.ts
@@ -1,7 +1,7 @@
 import { CoreCommand, Model, RemoteRevisionMessage } from "../../src";
 import { CommandSquisher, SquishedCoreCommand } from "../../src/collaborative/command_squisher";
+import { AutofillStore } from "../../src/components/autofill/autofill_store";
 import { toZone } from "../../src/helpers/zones";
-import { AutofillStore } from "../../src/plugins/ui_feature/autofill";
 import { MockTransportService } from "../__mocks__/transport_service";
 import { addRows, autofill, deleteRows, getCellContent, undo } from "../test_helpers";
 import { setCellContent } from "../test_helpers/commands_helpers";
