+++ tests/table/table_autofill_plugin.test.ts
@@ -1,8 +1,8 @@
 import { Model, UID } from "../../src";
+import { AutofillStore } from "../../src/components/autofill/autofill_store";
+import { TableAutofillStore } from "../../src/components/autofill/table_autofill_store";
 import { CellComposerStore } from "../../src/components/composer/composer/cell_composer_store";
 import { DEFAULT_TABLE_CONFIG } from "../../src/helpers/table_presets";
-import { AutofillStore } from "../../src/plugins/ui_feature/autofill";
-import { TableAutofillStore } from "../../src/plugins/ui_feature/table_autofill";
 import { DependencyContainer } from "../../src/store_engine/dependency_container";
 import {
   copy,
